app/database.py
from sqlalchemy import create_engine
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy.orm import sessionmaker
import os
from dotenv import load_dotenv

load_dotenv()

# Gunakan SQLite (file lokal) - untuk development
DATABASE_URL = os.getenv("DATABASE_URL", "sqlite:///./resep_nusantara.db")

# Untuk SQLite, perlu connect_args untuk multi-threading
engine = create_engine(
    DATABASE_URL,
    connect_args={"check_same_thread": False}  # Hanya untuk SQLite
)
SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)

Base = declarative_base()

# Dependency untuk mendapatkan session database
def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()